Implement task cards 0922–0941: contract and core consistency pass
contract package:
- Centralize all sentinel errors in errors.go; rename types for clarity
  (TypeValidation → "validation_failure", TypeTimeout → "timeout_failure")
- Remove WrapErrorf (near-twin of WrapError, no semantic distinction)
- Remove TenantID from Principal (use x/tenant/core context mechanism instead)
- Rename requestIDKey → requestIDContextKey for naming consistency
- Export ValidationErrors type for errors.As support
- applyValidationRule returns (*FieldError, error): programmer errors propagate directly
- Remove dead errOutOfRange check from DecodeRequestID
- BindJSON signature: (dst any, opts *BindOptions) replaces variadic …BindOptions
- Fix bindQuery to include ErrInvalidBindDst on pointer-to-non-struct path
- Update all BindJSON call sites to pass nil for default options

core package:
- Remove unused App.logError method
- Fix ServeHTTP handler_not_configured to use contract.CodeUnavailable
- Add named route shortcuts: GetWithName, PostWithName, PutWithName,
  DeleteWithName, PatchWithName, AnyWithName

x/tenant/core package:
- Create errors.go with all 7 sentinel errors centralized
- ConfigManager now embeds QuotaConfigProvider, PolicyConfigProvider,
  RateLimitConfigProvider; delete *FromConfig adapter types
- Rename InMemoryRateLimitProvider → InMemoryRateLimitManager
- Unexport Provider/Cache fields in ConfigPolicyEvaluator and
  CachedRoutePolicyProvider
- QuotaStore.Reserve/Release use QuotaReserveRequest/QuotaReleaseRequest structs
- All quota numeric types changed from int to int64
- Extract normalizeQuotaRequest helper; fix UTC in SlidingWindow
- Move Burst defaulting from Allow into RateLimitConfig.effectiveBurst()
- SlidingWindowQuotaManager returns ErrUnsupportedQuotaWindow for non-minute windows

Downstream fixes:
- x/tenant/config: cast legacy int DB columns to int64 for QuotaLimit
- x/tenant/quota: update tokens type and SetQuotaHeaders signature to int64
- x/tenant/transport: SetQuotaHeaders accepts int64
- x/tenant/resolve: remove principal-based tenant extraction (TenantID gone
  from Principal); remove DisablePrincipal option; update tests and examples

+++ x/tenant/core/hooks.go
@@ -39,10 +39,10 @@ type PolicyDecision struct {
 type QuotaDecision struct {
 	TenantID          string
 	Allowed           bool
-	Tokens            int
-	Requests          int
-	RemainingTokens   int
-	RemainingRequests int
+	Tokens            int64
+	Requests          int64
+	RemainingTokens   int64
+	RemainingRequests int64
 	RetryAfter        time.Duration
 	Status            int
 	// Err is non-nil when the decision was caused by a backend error rather than
